+++ resources/views/layouts/app.blade.php
@@ -40,9 +40,13 @@
 
     <x-footer />
 
+    <div id="toast-container"
+        class="fixed top-6 right-6 z-50 space-y-4">
+    </div>
+
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
     <script src="https://unpkg.com/lucide@latest"></script>
     <script>lucide.createIcons()</script>
-    
+
 </body>
 </html>